Moved collector specific configs to a new CollectorConfig class.
